[update]: link showCard and infoCard
