+++ src/components/Cards/Show/index.tsx
@@ -1,6 +1,7 @@
-import React from "react";
+import React, { useState } from "react";
 import './cardShow.css'
 import { InformationCircleIcon, HeartIcon } from "@heroicons/react/solid";
+import CardInfoPage from "../../../screens/card_info";
 
 export interface Idata {
   id: string
@@ -12,42 +13,49 @@ export interface Idata {
 
 export interface CardShowProp {
   showData: Idata[]
-  handleInfo: () => void
   handleFav: () => void
 }
 
-const CardShow = ({showData, handleInfo, handleFav}: CardShowProp) => {
+const CardShow = ({showData, handleFav}: CardShowProp) => {
+  const [openCard, setOpenCard] = useState(false)
+
   return (
     <ul className="card-show">
       {showData.map((showData) => (
-        <li
-          key={showData.id}
-          className={`card-poke ${showData.bgCard} transform hover:scale-110 motion-reduce:transform-none`}
-        >
-          <div>
-            <div className="-mt-px flex">
-              <button className="w-0 flex-1 flex p-2" onClick={handleInfo}>
-                <InformationCircleIcon className="h-7 w-7 text-blue-darkest hover:text-blue"/>
-              </button>
-              <button className="-ml-px w-0 flex-1 flex flex-row-reverse p-2" onClick={handleFav}>
-                <HeartIcon className={`h-7 w-7 ${showData.isFav? 'text-red-dark hover:text-red' : ''}`}/>
-              </button>
+        <>
+          {
+            openCard &&
+            <CardInfoPage pokemonName={showData.name} isOpen={openCard} handleClose={() => setOpenCard(false)} />
+          }
+          <li
+            key={showData.id}
+            className={`card-poke ${showData.bgCard} transform hover:scale-110 motion-reduce:transform-none`}
+          >
+            <div>
+              <div className="-mt-px flex">
+                <button className="w-0 flex-1 flex p-2" onClick={() => setOpenCard(true)}>
+                  <InformationCircleIcon className="h-7 w-7 text-blue-darkest hover:text-blue"/>
+                </button>
+                <button className="-ml-px w-0 flex-1 flex flex-row-reverse p-2" onClick={handleFav}>
+                  <HeartIcon className={`h-7 w-7 ${showData.isFav? 'text-red-dark hover:text-red' : ''}`}/>
+                </button>
+              </div>
             </div>
-          </div>
-          <div className="flex-1 flex flex-col p-7">
-            <img
-              className="w-36 h-36 flex-shrink-0 mx-auto bg-black rounded-md"
-              src={showData.imageUrl}
-              alt=""
-            />
-            <h3 className="mt-6 text-gray-900 font-press-start text-sm font-medium">
-              {showData.name}
-            </h3>
-            <h3 className="mt-4 text-gray-900 font-quicksand text-base font-medium">
-              {showData.name}
-            </h3>
-          </div>
-        </li>
+            <div className="flex-1 flex flex-col p-7">
+              <img
+                className="w-36 h-36 flex-shrink-0 mx-auto bg-black rounded-md"
+                src={showData.imageUrl}
+                alt=""
+              />
+              <h3 className="mt-6 text-gray-900 font-press-start text-sm font-medium">
+                {showData.name}
+              </h3>
+              <h3 className="mt-4 text-gray-900 font-quicksand text-base font-medium">
+                {showData.name}
+              </h3>
+            </div>
+          </li>
+        </>      
       ))}
     </ul>
   );
